In this part I first wanted to treat the problem as an anomaly detection problem, and training models on REAL data, for this I used different algorithms of this class such as one class SVM, isolation forest or robust variance.
However it only produced 0.74896. And this problem here is very specific, it is not just like ai detection content, both texts have been modified so I chose to test a binary SVM to take advantage of this, as we have two classes: FAKE and REAL.

The detection is based on mean of embeddings for a certain layer so each text is assinged to one vectors.